requirements.txt
snowflake-snowpark-python==1.32.0
snowflake-connector-python==3.15.0
PyYAML==6.0.2